requirements.txt
anyio==4.8.0
appdirs==1.4.4
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.2.0
dateparser==1.2.1
demjson3==3.0.6
ensembledata==0.2.4
facebook-scraper==0.2.59
fake-useragent==2.0.3
greenlet==3.1.1
groq
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.1
parse==1.20.2
playwright==1.50.0
pyee==12.1.1
pyppeteer==0.0.25
pyquery==2.0.1
pyTelegramBotAPI
python-dateutil==2.9.0.post0
python-dotenv
pytubefix==8.12.2
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tiktokapi==7.0.0
tqdm==4.67.1
typing-extensions==4.12.2
tzlocal==5.3
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
